requirements.txt
llama-cpp-python==0.2.56
sentence-transformers==2.7.0
chromadb==0.5.3
langchain==0.2.6
langchain-community==0.2.6
langgraph==0.1.19
lxml==5.2.2
pymupdf==1.24.5
beautifulsoup4==4.12.3
httpx==0.27.0
python-can==4.4.2
can-isotp==2.0.2
fastapi==0.111.0
uvicorn==0.30.1
gradio==4.37.2
typer==0.12.3
pytest==8.2.2
pytest-asyncio==0.23.7
jinja2==3.1.4
watchdog==4.0.1
python-dotenv==1.0.1
pydantic==2.7.4
rich==13.7.1
